Describe the Protocol
The Stack Tokens protocol is a simple way to utilize Bitcoin's Script stack as a mechanism for storing arbitrary application data.

Here we describe the protocol, including how we can identify that there is a Stack Token, how to allow for future versions, and how to identify which application owns this Stack Token.